is_to_vfat_migration: initial commit of scripts
This patch provides a couple of scripts to facilitate ISO to VFAT
migration.  There are a handful of obstacles that need to be addressed,
mostly surrounding the fact that libvirt/qemu don't support IDE drive
live migration.  As such, we need to copy the contents of the ISO that
is used for config drive to a vfat (virtual) device, unmount the CD
drive, copy the new vfat device to ceph, and mount the new vfat drive.
A reboot is required to fully remove the IDE CD drive.

This is all accomplished by calling the iso_to_vfat.sh script and
passing in the virsh instance/domain name (usually instance-000000).