Initial commit of nova_audit_routers.py
This script checks for routers that neutron knows about and compares
that to VMs that belong to the service tenant (of which should only have
router VMs).  It's concievable that a VM is created for a router, but
then neutron deletes the router, but not the supporting VM, thus
creating an orphaned router VM.  This script should help track those
down.